+++ currency-conversion-api/readme.md
@@ -1,36 +1,72 @@
-<h2 align="center">RestTemplate</h2>
-<ol>
-	<li>RestTemplate is a wrapper which provide Http methods</li>
-	<li>here we used getEntity, give url, response type and url map</li>
-</ol>
+Dependencies
+============
 
-<h2 align="center">Fiegn Rest Client</h2>
-<ol>
-	<li>Feign is a declarative web service client.</li>
-	<li>add @EnableFeignClients at application level</li>
-	<li>Create an interface and annotate it with @FeignClient</li>
-	<li>Now give two arguments</li>
-	<ol>
-		<li>name = "the name of of the service, which you want to talk"</li>
-		<li>url = "localhost:8000" the port number</li>
-	</ol>
-	<li>Now: create a any mapping for this service</li>
-</ol>
-
-<h2 align="center">Netflix-Ribbon</h2>
-<ol>
-	<li>@RibbonClient(name="name of service which you want to talk")</li>
-	<li>Remove the url from @FiegnClient</li>
-	<li>Add `name of the service which you want to talk from this service`.ribbon.listofServers=http://localhost:`port of service which you want to talk', more and, more</li>
-</ol>
+- #### Web (embedded tomcat + MVC)
+    - spring-boot-starter-web
+- #### Dev Tools
+    - spring-boot-devtools
+    - lombok
+    - spring-boot-configuration-processor
+- #### Dev Ops
+    - spring-boot-starter-actuator
+- #### Cloud
+    - ##### spring-cloud-starter-netflix-eureka-client
+    - ##### spring-cloud-starter-netflix-ribbon
+    - ##### spring-cloud-starter-openfeign
+    - ##### spring-cloud-starter-zipkin
+- #### Swagger
+    - springfox-swagger2
+    - springfox-swagger-ui
+- #### Test
+    - spring-boot-starter-test
 
 
-<h2 align="center">Calling currency-exchange-service from API gateway </h2>
-<ol>
-	<li>We will change the @FeignClient name param to API gateway</li>
-	<li>As well change the getMapping in FiegnProxy and append the application name (targeted service name)</li>
-	<li>our API gateway talk the eureka naming server and get the service which is mentioned</li>
-	<li>Simply call this service, then this service call api gate-way server to call the currency exchange service</li>
-	<li>To call this service from the api gateway server, simply change the port and append service name</li>
-	<li>http://localhost:8755/currency-conversion-service/currency-conversion-feign/from/USD/to/PKR/quantity/1000</li>
-</ol>
+Feign Rest Client
+----------------
+- ### About
+    - Declarative Web Service Client
+- ### Config
+    - Add @EnableFeignClients at Application Level
+    - Create an interface and annotate it with ```@FeignClient```
+    - ##### Config (In case of no gateway-server)
+        - ```@FeignClient``` gets two arguments 
+            -   ```name``` the name of the service to which we want to talk
+            -   ```url``` the url of the service
+    - ##### Config (With gateway-server and registered with naming server)
+        - ```@FeignClient``` gets one argument
+            - ```name``` the name of api-gateway-server
+    - Add request mapping inside this interface (prefixed with the name of service in case of gateway-server)   
+    
+Netflix-Ribbon
+--------------
+- Netflix-Ribbon is a loadbalancer, and it can work along with feign Client
+- ##### Config
+    - Add ```@RibbonClient``` on the ```@FiegnClient``` annotated interface
+    - In the arguments of ```@RibbonClient``` put the name of service to which you want to talk
+
+Rest Template
+-------------
+- RestTemplate is a wrapper which provide Http methods
+
+How its work
+------------
+- We have
+    - Eureka Naming Server
+    - Zuul Api gateway Server
+    - Feign Client 
+- When we call a service from feign clients, then our request goes to API gateway server, it talks with eureka naming server and get the service which is being called.
+- We can simply directly call a service from our API gateway just prepending the name of being called service and change port to API gateway.
+
+Distributed Tracing
+-------------------
+1.  Add dependency ```spring-cloud-starter-zipkin```
+2. It compiles two dependencies
+    1. ```spring-cloud-starter-sleuth```
+    2. ```spring-cloud-sleuth-zipkin```
+3.  Sleuth will assign a unique id to each request
+4.  Sleuth-Zipkin create Zipkin-compatible traces via HTTP
+5.  Create a bean which return ```Sampler.ALWAYS_SAMPLE```
+6.  Zipkin needs a message broker and by default it is ```rabbit-mq```
+7.  We have to run rabbit-mq and zipkin-server and  over-ride two properties
+    1. ```spring.rabbitmq.addresses=amqp://localhost:5672/```
+    2. ```spring.zipkin.base-url=http://localhost:9411/```
