+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/CurrencyConversionClientApplication.java
@@ -1,4 +1,4 @@
-package com.ubaid.ms.currencyconversionclient;
+package com.ubaid.ms.currencyconversionapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
