+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/controller/ExchangeController.java
@@ -1,17 +1,13 @@
-package com.ubaid.ms.currencyconversionclient.controller;
+package com.ubaid.ms.currencyconversionapi.controller;
 
-import com.netflix.zuul.context.RequestContext;
-import com.ubaid.ms.currencyconversionclient.dto.ExchangeValueDTO;
-import com.ubaid.ms.currencyconversionclient.service.CurrencyConversionService;
+import com.ubaid.ms.currencyconversionapi.dto.ExchangeValueDTO;
+import com.ubaid.ms.currencyconversionapi.service.CurrencyConversionService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.HttpServletResponse;
-
 
 @RestController
 @RequestMapping(value = "exchange")
@@ -30,10 +26,7 @@ public class ExchangeController {
             @PathVariable("from") String from,
             @PathVariable("to") String to,
             @PathVariable("quantity") Integer quantity) {
-        ExchangeValueDTO currencyConversion = currencyConversionService.convertCurrency(from, to, quantity);
-        HttpServletResponse response = RequestContext.getCurrentContext().getResponse();
-        log.info("response =========> {}", response);
-        return currencyConversion;
+        return currencyConversionService.convertCurrency(from, to, quantity);
     }
 
 }
