+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/controller/LimitController.java
@@ -1,16 +1,13 @@
-package com.ubaid.ms.currencyconversionclient.controller;
+package com.ubaid.ms.currencyconversionapi.controller;
 
 
-import com.netflix.zuul.context.RequestContext;
-import com.ubaid.ms.currencyconversionclient.entity.Limit;
-import com.ubaid.ms.currencyconversionclient.feignProxy.LimitProxy;
+import com.ubaid.ms.currencyconversionapi.entity.Limit;
+import com.ubaid.ms.currencyconversionapi.feignProxy.LimitProxy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.HttpServletResponse;
 
 @RestController
 @RequestMapping(value = "splgbi")
@@ -26,9 +23,6 @@ public class LimitController {
 
     @GetMapping("/limits")
     public Limit getCurrencyConversion() {
-        Limit limitConfig = proxy.getLimitConfig();
-        HttpServletResponse response = RequestContext.getCurrentContext().getResponse();
-        log.info("response =========> {}", response);
-        return limitConfig;
+        return proxy.getLimitConfig();
     }
 }
