+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/feignProxy/CurrencyConversionServiceProxy.java
@@ -1,0 +1,15 @@
+package com.ubaid.ms.currencyconversionapi.feignProxy;
+
+import com.ubaid.ms.currencyconversionapi.dto.ConvertedCurrency;
+import org.springframework.cloud.netflix.ribbon.RibbonClient;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+@FeignClient(name = "netflix-zuul-api-gateway-server")
+@RibbonClient(name = "currency-conversion-service")
+public interface CurrencyConversionServiceProxy {
+
+    @GetMapping("currency-conversion-service/currency-conversion/currency/{currency}/rate/{conversion-rate}")
+    ConvertedCurrency convert(@PathVariable("currency") Integer currency, @PathVariable("conversion-rate") Integer rate);
+}
