+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/feignProxy/LimitProxy.java
@@ -1,6 +1,6 @@
-package com.ubaid.ms.currencyconversionclient.feignProxy;
+package com.ubaid.ms.currencyconversionapi.feignProxy;
 
-import com.ubaid.ms.currencyconversionclient.entity.Limit;
+import com.ubaid.ms.currencyconversionapi.entity.Limit;
 import org.springframework.cloud.netflix.ribbon.RibbonClient;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
