+++ currency-conversion-api/src/main/java/com/ubaid/ms/currencyconversionapi/service/CurrencyConversionService.java
@@ -1,6 +1,6 @@
-package com.ubaid.ms.currencyconversionclient.service;
+package com.ubaid.ms.currencyconversionapi.service;
 
-import com.ubaid.ms.currencyconversionclient.dto.ExchangeValueDTO;
+import com.ubaid.ms.currencyconversionapi.dto.ExchangeValueDTO;
 
 public interface CurrencyConversionService {
     ExchangeValueDTO convertCurrency(String fromCurrency, String toCurrency, Integer quantity);
